Revert "Merge pull request #44 from gpsubprojectfirst/hy"
This reverts commit acf6fec8abf493e66204f1e9f27d31b244a52461, reversing
changes made to c64e282cbabf9bafb5988796aba7d0e142a1d81f.

+++ 0806_임시프로젝트파일_ 하영/GP_subproject_first/Build.h
@@ -19,6 +19,6 @@ public:
 
 	virtual void Rest(float Delta) override;
 
-	void CalcDirection(float xvec, float yvec) override;
+	void CalcDirection(int xvec, int yvec) override;
 	std::vector<Gdiplus::Rect> restRc;
 };
